+++ logs/DOC_LOG.md
@@ -519,3 +519,4 @@ executados.
 >>>>>>> origin/codex/criar-fluxo-unificado-para-gerenciar-link-de-pagamento
 ## [2025-07-02] Inclusao de pbRetry nas rotas principais e docs atualizadas. - Lint e Build executados.
 ## [2025-08-16] Ajustada funcao calculateGross comparando valor de credito com Pix e testes atualizados. Lint e build executados.
+## [2025-07-02] Implementado webhook assíncrono com fila e worker. Lint e build executados.
